Move funcitons into files that make more sense
